Fixes for the Linux build.

+++ src/builtin_words.cpp
@@ -425,7 +425,7 @@ namespace sorth
         // unregister it now.
 
         at_exit_interpreter = nullptr;
-        at_exit_value = 0;
+        at_exit_value = (int64_t)0;
 
         interpreter->release_context();
         interpreter->clear_stack();
@@ -527,7 +527,7 @@ namespace sorth
         insert_user_instruction(interpreter,
             {
                 .id = OperationCode::Id::read_variable,
-                .value = 0
+                .value = (int64_t)0
             });
     }
 
@@ -537,7 +537,7 @@ namespace sorth
         insert_user_instruction(interpreter,
             {
                 .id = OperationCode::Id::write_variable,
-                .value = 0
+                .value = (int64_t)0
             });
     }
 
@@ -577,7 +577,7 @@ namespace sorth
         insert_user_instruction(interpreter,
             {
                 .id = OperationCode::Id::unmark_loop_exit,
-                .value = 0
+                .value = (int64_t)0
             });
     }
 
@@ -597,7 +597,7 @@ namespace sorth
         insert_user_instruction(interpreter,
             {
                 .id = OperationCode::Id::unmark_catch,
-                .value = 0
+                .value = (int64_t)0
             });
     }
 
@@ -637,7 +637,7 @@ namespace sorth
         insert_user_instruction(interpreter,
             {
                 .id = OperationCode::Id::jump_loop_start,
-                .value = 0
+                .value = (int64_t)0
             });
     }
 
@@ -647,7 +647,7 @@ namespace sorth
         insert_user_instruction(interpreter,
             {
                 .id = OperationCode::Id::jump_loop_exit,
-                .value = 0
+                .value = (int64_t)0
             });
     }
 
@@ -729,7 +729,7 @@ namespace sorth
             else if (top_code[i].id == OperationCode::Id::jump_target)
             {
                 jump_targets.insert({ as_string(interpreter, top_code[i].value), i });
-                top_code[i].value = 0;
+                top_code[i].value = (int64_t)0;
             }
         }
 
@@ -1161,7 +1161,7 @@ namespace sorth
         }
         else
         {
-            interpreter->push(std::strtoll(string.c_str(), nullptr, 10));
+            interpreter->push((int64_t)std::strtoll(string.c_str(), nullptr, 10));
         }
     }
 
@@ -1212,7 +1212,7 @@ namespace sorth
 
                     interpreter->push(",");
                     interpreter->push(";");
-                    interpreter->push(2);
+                    interpreter->push((int64_t)2);
                     word_code_compile_until_words(interpreter);
                     auto found_word = as_string(interpreter, interpreter->pop());
                     --current_token;
@@ -1234,7 +1234,7 @@ namespace sorth
                 }
                 else
                 {
-                    definition_ptr->defaults.push_back(0);
+                    definition_ptr->defaults.push_back((int64_t)0);
                 }
             }
         }
@@ -1753,13 +1753,13 @@ namespace sorth
 
     void word_exit_success(InterpreterPtr& interpreter)
     {
-        interpreter->push(EXIT_SUCCESS);
+        interpreter->push((int64_t)EXIT_SUCCESS);
     }
 
 
     void word_exit_failure(InterpreterPtr& interpreter)
     {
-        interpreter->push(EXIT_FAILURE);
+        interpreter->push((int64_t)EXIT_FAILURE);
     }
